Switch to Blosc2.

+++ src/HDF5.NET/Filters/ShuffleGeneric.cs
@@ -6,13 +6,13 @@ namespace HDF5.NET
     public static class ShuffleGeneric
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static unsafe void Shuffle(ulong type_size, ulong vectorizable_blocksize, ulong blocksize, byte* source, byte* dest)
+        public static unsafe void Shuffle(int type_size, int vectorizable_blocksize, int blocksize, byte* source, byte* dest)
         {
             throw new NotImplementedException();
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static unsafe void Unshuffle(ulong type_size, ulong vectorizable_blocksize, ulong blocksize, byte* source, byte* dest)
+        public static unsafe void Unshuffle(int type_size, int vectorizable_blocksize, int blocksize, byte* source, byte* dest)
         {
             throw new NotImplementedException();
         }
